Get One Image feature
* new command added to get an image
* replaced old command with `start_posting`
* Pulled out main loop function to post an image and made it stand alone
* Pulled other utility functions into utility.py
* Update to readme removing unneeded sections
