+++ .gitignore
@@ -127,3 +127,4 @@ dmypy.json
 images/
 guilds/
 .vscode
+.idea
