+++ bot.py
@@ -10,6 +10,8 @@ from dotenv import load_dotenv
 
 from bot.main_loop import main_loop
 from bot.startup import check_running
+from bot.get_image import send_image
+from bot.utility import file_to_array
 
 dotenv_path = Path('.env')
 load_dotenv(dotenv_path=dotenv_path)
@@ -18,7 +20,7 @@ TOKEN = os.getenv('DISCORD_TOKEN')
 bot = commands.Bot(command_prefix='!')
 
 
-@bot.command(name='uWu')
+@bot.command(name='start_posting')
 async def start_posting(ctx):
     guild_id = ctx.channel.guild.id
     channel_id = ctx.channel.id
@@ -36,6 +38,13 @@ async def start_posting(ctx):
 
     await ctx.send('(o･｀Д´･o) Baka!')
     print(f'Ending posting for {guild_id}')
+
+
+@bot.command(name='!uWu!!')
+async def get_one(ctx):
+    guild_id = ctx.channel.guild.id
+    await send_image(
+        ctx, file_to_array(f'guilds/images_{guild_id}.txt'), guild_id)
 
 
 @bot.event
